Temporary fix on final compiling error
This was the last compiling error I could find, for the different tests.

The files testVoronoiDensityGrid.exe and testHydroIntegrator.exe would fail to compile, as both showed the error "C2034 -  'ExactGeometricTests::<unnamed-tag>::mantissa': type of bit field too small for number of bits".

It seems to be having an issue with compiling (at least on my version of VisualStudio) the ExactGeometricTests.hpp file when the mantissa is declared as having anything more than 32 bits. This is an issue, as the comments mention that the IEEE 754 standard specifies the mantissa as having 52 bit precision.

It compiles when I change it from 52 to 32, but I know that this is wrong, so it is only a temporary fix.