Add missing jfac normalization in IonizationStateCalculator
As pointed out by Lewis, a jfac normalisation was missing for j_metals in
IonizationStateCalculator::calculate_ionization_state(). This only matters when j_metals is used to calculate
the ionization state of metals when the temperature is fixed, which is not a scenario that was ever used for
scientific simulations (except for the early iterations, when it does not affect the final result). Verified
that this makes a difference by running the LexingtonHII20 benchmark with a fixed temperature.

Also replaced a now deprecated argument to Axes.set_yscale() in the Lexington plotting script.